Created a new check.html that supports using the canvas as a clickable link.
User wanted to add clickable links to the canvas. After investigating and figuring out the correct absolute to relative positioning I now can assign a URL to the canvas. Clicking on the canvas allows the user to be directed to a new window with th referenced URL being broughtup. I've added tiny blue 10x10 boxes to each canvas to give a visual. Since this doesn't seem to mean much I may remove these in a later version. This was the reason for an updated ip-results.json. Currently the ping-check.sh and the ip-addr.txt don't have support for the URL. This will be added later.

TODO: There are a few FIXME and TODO tags that will need to be addressed in the check.html file. At the moment I don't deal with a window resize. Also some more code clean up and more comments on why I did what I did. The whole absolute vs relative position was a bit confusing as you had to reference the canvas after it was added to the page. Also I found that Firefox has some pretty strange behaviours when it hits an error (like suddenly breaking out of the function with no error messages and try {} catch(e) {} doesn't do anything. I'm guessing this is some kind of memory leak when debug pounding on the browser.

[ticket: X]