Initial History file
My git commits are not all that great so I'm trying to add some kind of a blog (a History) file which I can record what generall changes were made.

I'm used to having a nice long list of changes in the file that mostly explain what changes were made. So far I'm not there with git.